角色设计文档完成；游戏开始结束输入模块切换完成；设置界面数据保存和按钮功能完成

+++ Assets/Game/Scripts/ScriptableObject/CharacterData_SO.cs
@@ -12,5 +12,5 @@ public class CharacterData_SO : ScriptableObject
     public float attackSpeedMul = 1.0f;
     public float criticalRate = 0.0f;
     public int attackRange = 450;
-    public int moveSpeed = 10;
+    public float moveSpeed = 10;
 }
